# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
  /client/history:
    get:
      operationId: getClientHistory
      tags: [appointments, orders]
      summary: Upcoming appointments and issued receipts belonging to the verified client
      responses:
        "200":
          description: Client history
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ClientHistory" }
        "401":
          description: Verified client sign-in required
  /client/receipts/{id}/view:
    get:
      operationId: viewClientReceipt
      tags: [orders]
      summary: View a receipt owned by the verified client
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Receipt PDF
          content:
            application/pdf:
              schema: { type: string, format: binary }
        "404":
          description: Receipt not found
  /client/receipts/{id}/pdf:
    get:
      operationId: downloadClientReceipt
      tags: [orders]
      summary: Download a receipt owned by the verified client
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Receipt PDF
          content:
            application/pdf:
              schema: { type: string, format: binary }
        "404":
          description: Receipt not found
  /admin/orders:
    get:
      operationId: listWigOrders
      tags: [orders]
      summary: List private wig orders
      responses:
        "200":
          description: Orders
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/WigOrder" }
    post:
      operationId: createWigOrder
      tags: [orders]
      summary: Create a wig order
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/WigOrderInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WigOrder" }
  /admin/orders/{id}:
    patch:
      operationId: updateWigOrder
      tags: [orders]
      summary: Update an order or payment total
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/WigOrderInput" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WigOrder" }
  /admin/orders/{id}/receipts:
    get:
      operationId: listWigReceipts
      tags: [orders]
      summary: List previously issued receipt snapshots
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Receipts
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/WigReceipt" }
    post:
      operationId: issueWigReceipt
      tags: [orders]
      summary: Issue a receipt from the saved order
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "201":
          description: Issued
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WigReceipt" }
  /admin/orders/{id}/receipts/{receiptId}/pdf:
    get:
      operationId: downloadWigReceipt
      tags: [orders]
      summary: Download protected receipt PDF
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
        - in: path
          name: receiptId
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: PDF document
          content:
            application/pdf:
              schema: { type: string, format: binary }
  /admin/orders/{id}/receipts/{receiptId}/view:
    get:
      operationId: viewWigReceipt
      tags: [orders]
      summary: View protected receipt PDF in the browser
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
        - in: path
          name: receiptId
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Inline PDF document
          content:
            application/pdf:
              schema: { type: string, format: binary }
...
components:
  schemas:
    ClientHistory:
      type: object
      required: [email, appointments, receipts]
      properties:
        email: { type: string }
        appointments:
          type: array
          items: { $ref: "#/components/schemas/ClientAppointment" }
        receipts:
          type: array
          items: { $ref: "#/components/schemas/WigReceipt" }
    ClientAppointment:
      type: object
      required: [id, service, appointmentDate, appointmentTime, status]
      properties:
        id: { type: integer }
        service: { type: string }
        appointmentDate: { type: string, format: date }
        appointmentTime: { type: string }
        status: { $ref: "#/components/schemas/AppointmentStatus" }
    WigOrderInput:
      type: object
      additionalProperties: false
      required: [kind, itemCode, orderDate, customerName, priceCents, taxRateMilliPercent, amountPaidCents]
      properties:
        kind: { type: string, enum: [stock, custom] }
        itemCode: { type: string, minLength: 1, maxLength: 100 }
        orderDate: { type: string, pattern: '^\d{4}-\d{2}-\d{2}$' }
        customerName: { type: string, minLength: 1, maxLength: 200 }
        phone: { type: string, maxLength: 100 }
        email: { type: string, maxLength: 200 }
        notes: { type: string, maxLength: 3000 }
        style: { type: string, maxLength: 200 }
        capSize: { type: string, maxLength: 200 }
        lengthInch: { type: string, maxLength: 200 }
        hairType: { type: string, maxLength: 200 }
        part: { type: string, maxLength: 200 }
        layers: { type: string, maxLength: 200 }
        density: { type: string, maxLength: 200 }
        color: { type: string, maxLength: 200 }
        highlights: { type: string, maxLength: 500 }
        priceCents: { type: integer, minimum: 0 }
        taxRateMilliPercent: { type: integer, minimum: 0, maximum: 100000 }
        amountPaidCents: { type: integer, minimum: 0 }
    WigOrder:
      type: object
      required: [id, kind, itemCode, orderDate, customerName, phone, email, notes, style, capSize, lengthInch, hairType, part, layers, density, color, highlights, priceCents, taxRateMilliPercent, amountPaidCents, taxCents, tariffCents, totalCents, amountDueCents, needsReview, reviewIssues, sourceSheet, sourceRowNumber, sourceValues, createdAt]
      properties:
        id: { type: integer }
        kind: { type: string, enum: [stock, custom] }
        itemCode: { type: string, nullable: true }
        orderDate: { type: string, nullable: true }
        customerName: { type: string, nullable: true }
        phone: { type: string }
        email: { type: string }
        notes: { type: string }
        style: { type: string }
        capSize: { type: string }
        lengthInch: { type: string }
        hairType: { type: string }
        part: { type: string }
        layers: { type: string }
        density: { type: string }
        color: { type: string }
        highlights: { type: string }
        priceCents: { type: integer, nullable: true }
        taxRateMilliPercent: { type: integer }
        amountPaidCents: { type: integer, nullable: true }
        taxCents: { type: integer, nullable: true }
        tariffCents: { type: integer, minimum: 0 }
        totalCents: { type: integer, nullable: true }
        amountDueCents: { type: integer, nullable: true }
        needsReview: { type: boolean }
        reviewIssues:
          type: array
          items: { type: string }
        sourceSheet: { type: string, nullable: true }
        sourceRowNumber: { type: integer, nullable: true }
        sourceValues:
          type: object
          nullable: true
          additionalProperties: { type: string }
        createdAt: { type: string, format: date-time }
    WigReceipt:
      type: object
      required: [id, orderId, issuedAt, receiptNumber, customerName, itemCode]
      properties:
        id: { type: integer }
        orderId: { type: integer }
        issuedAt: { type: string, format: date-time }
        receiptNumber: { type: string }
        customerName: { type: string, nullable: true }
        itemCode: { type: string, nullable: true }
...
    AppointmentStatus:
      type: string
      enum: [pending, confirmed, completed, cancelled]
...