# lib/api-zod/src/generated/types/clientAppointment.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AppointmentStatus } from './appointmentStatus';

export interface ClientAppointment {
  id: number;
  service: string;
  appointmentDate: Date;
  appointmentTime: string;
  status: AppointmentStatus;
}