# lib/api-zod/src/generated/types/clientHistory.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ClientAppointment } from './clientAppointment';
import type { WigReceipt } from './wigReceipt';

export interface ClientHistory {
  email: string;
  appointments: ClientAppointment[];
  receipts: WigReceipt[];
}